Added Role Seeding

+++ src/FamilyHubs.Referral.Data/Entities/Role.cs
@@ -1,6 +1,6 @@
 ﻿namespace FamilyHubs.Referral.Data.Entities;
 
-public class Role : EntityBase<long>
+public class Role : EntityBase<byte>
 {
     public required string Name { get; set; }
     public string? Description { get; set; }
